[Refactor] move the common codes into a new file `recentOpen.ts`
